refactor(repl): split session/builtins.rs into capabilities/batched/authoring
builtins.rs had grown to ~1175 lines covering five jobs (engine
construction, single-shot capability calls, batched variants, graph
authoring, and shared helpers). Converted to a builtins/ module
directory and split mechanically into capabilities.rs (model_query/
tool_call/agent_query/graph_run), batched.rs (the *_batched variants),
and authoring.rs (graph_define/validate/compile/diff/register), with
mod.rs keeping engine construction and shared helpers. External path
(session::builtins) unchanged. No behavior change.
