+++ src/repl/session/builtins/mod.rs
@@ -342,610 +342,13 @@ fn model_value(text: String, finish_reason: Option<String>, structured: bool) ->
     }
 }
 
-// ── Single capability implementations ───────────────────────────────────────
+mod authoring;
+mod batched;
+mod capabilities;
 
-fn model_query_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    params: &Map,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let model_name =
-        map_str(params, "model").ok_or_else(|| invalid(ctx, "model_query: missing `model`"))?;
-    bump_model(ctx)?;
-    let model = ctx
-        .registry
-        .model(&model_name)
-        .ok_or_else(|| raise(ctx, TinyAgentsError::ModelNotFound(model_name.clone())))?;
-    let request = build_model_request(&model_name, params);
-    let start = Instant::now();
-    let response = bridge_block_on(ctx.buffers.deadline(), model.invoke(&ctx.state, request))
-        .map_err(|err| raise(ctx, err))?;
-    let elapsed = start.elapsed();
-    let finish_reason = response.finish_reason.clone();
-    let text = Message::Assistant(response.message).text();
-    record(
-        ctx,
-        ReplCallKind::Model,
-        &model_name,
-        json!({ "chars": text.len() }),
-        elapsed,
-    );
-    Ok(model_value(
-        text,
-        finish_reason,
-        map_bool(params, "structured").unwrap_or(false),
-    ))
-}
-
-fn tool_call_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    params: &Map,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let tool_name =
-        map_str(params, "tool").ok_or_else(|| invalid(ctx, "tool_call: missing `tool`"))?;
-    bump_tool(ctx)?;
-    let tool = ctx
-        .registry
-        .tool(&tool_name)
-        .ok_or_else(|| raise(ctx, TinyAgentsError::ToolNotFound(tool_name.clone())))?;
-    let arguments = map_json(params, "arguments").unwrap_or(Value::Null);
-    let call = ToolCall {
-        id: new_call_id().as_str().to_string(),
-        name: tool_name.clone(),
-        arguments: arguments.clone(),
-    };
-    let start = Instant::now();
-    let result = bridge_block_on(ctx.buffers.deadline(), tool.call(&ctx.state, call))
-        .map_err(|err| raise(ctx, err))?;
-    let elapsed = start.elapsed();
-    record(
-        ctx,
-        ReplCallKind::Tool,
-        &tool_name,
-        json!({ "arguments": arguments }),
-        elapsed,
-    );
-    if let Some(error) = result.error {
-        return Err(raise(ctx, TinyAgentsError::Tool(error)));
-    }
-    let structured = map_bool(params, "structured").unwrap_or(false);
-    if structured && result.raw.is_some() {
-        let mut map = Map::new();
-        map.insert("content".into(), Dynamic::from(result.content));
-        map.insert(
-            "raw".into(),
-            repl_value_to_dynamic(&json_to_repl_value(&result.raw.unwrap_or(Value::Null))),
-        );
-        Ok(Dynamic::from_map(map))
-    } else {
-        Ok(Dynamic::from(result.content))
-    }
-}
-
-fn agent_query_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    params: &Map,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    use crate::graph::subagent_node::SubAgentInput;
-    let agent_name =
-        map_str(params, "agent").ok_or_else(|| invalid(ctx, "agent_query: missing `agent`"))?;
-    bump_agent(ctx)?;
-    check_depth(ctx)?;
-    let agent = ctx.registry.agent(&agent_name).ok_or_else(|| {
-        raise(
-            ctx,
-            TinyAgentsError::Capability(format!("agent `{agent_name}` is not registered")),
-        )
-    })?;
-    let prompt = map_str(params, "prompt")
-        .or_else(|| map_str(params, "input"))
-        .unwrap_or_default();
-    let mut input = SubAgentInput::prompt(prompt);
-    if let Some(data) = map_json(params, "input") {
-        input = input.with_data(data);
-    }
-    let start = Instant::now();
-    let output = bridge_block_on(ctx.buffers.deadline(), agent.run(input, ctx.events.clone()))
-        .map_err(|err| raise(ctx, err))?;
-    record(
-        ctx,
-        ReplCallKind::Agent,
-        &agent_name,
-        json!({ "model_calls": output.model_calls, "tool_calls": output.tool_calls }),
-        start.elapsed(),
-    );
-    Ok(Dynamic::from(output.text))
-}
-
-/// Resolves a registered graph blueprint and records the run, returning a
-/// reference to the resolved topology.
-///
-/// Resolving a registered graph routes through the capability registry; the
-/// REPL hands back the resolved blueprint reference (graph id, start node, node
-/// count) rather than installing or stepping topology here. Materializing a
-/// `CompiledGraph` and driving its super-steps is owned by the graph runtime
-/// and wired in a later slice; this keeps the REPL an orchestration surface,
-/// not a topology-mutation surface.
-fn graph_run_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    params: &Map,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let graph_name =
-        map_str(params, "graph").ok_or_else(|| invalid(ctx, "graph_run: missing `graph`"))?;
-    bump_graph(ctx)?;
-    check_depth(ctx)?;
-    let blueprint = ctx
-        .registry
-        .graph_blueprint(&graph_name)
-        .ok_or_else(|| {
-            raise(
-                ctx,
-                TinyAgentsError::Capability(format!("graph `{graph_name}` is not registered")),
-            )
-        })?
-        .clone();
-    record(
-        ctx,
-        ReplCallKind::Graph,
-        &graph_name,
-        json!({ "nodes": blueprint.nodes.len() }),
-        Duration::default(),
-    );
-    Ok(blueprint_reference(&blueprint))
-}
-
-/// Builds the script-visible reference map for a resolved graph blueprint.
-fn blueprint_reference(blueprint: &Blueprint) -> Dynamic {
-    let mut map = Map::new();
-    map.insert("graph".into(), Dynamic::from(blueprint.graph_id.clone()));
-    map.insert("start".into(), Dynamic::from(blueprint.start.clone()));
-    map.insert("nodes".into(), Dynamic::from(blueprint.nodes.len() as i64));
-    map.insert("resolved".into(), Dynamic::from(true));
-    Dynamic::from_map(map)
-}
-
-// ── Graph-authoring implementations ─────────────────────────────────────────
-
-fn graph_define_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    params: &Map,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let name =
-        map_str(params, "name").ok_or_else(|| invalid(ctx, "graph_define: missing `name`"))?;
-    let source =
-        map_str(params, "source").ok_or_else(|| invalid(ctx, "graph_define: missing `source`"))?;
-
-    // Check the limit up front (without consuming a slot) so a session that
-    // has already hit the cap fails fast instead of paying for a parse and
-    // compile it can't keep the result of anyway.
-    if ctx.counters.lock().expect("counters poisoned").graph_def >= ctx.policy.max_graph_definitions
-    {
-        return Err(raise(
-            ctx,
-            TinyAgentsError::LimitExceeded(format!(
-                "graph definition limit ({}) exceeded",
-                ctx.policy.max_graph_definitions
-            )),
-        ));
-    }
-    if source.len() > ctx.policy.max_script_bytes {
-        return Err(raise(
-            ctx,
-            TinyAgentsError::LimitExceeded(format!(
-                "graph source is {} bytes, exceeding max_script_bytes ({})",
-                source.len(),
-                ctx.policy.max_script_bytes
-            )),
-        ));
-    }
-
-    let label = ctx
-        .language
-        .as_ref()
-        .map(|l| l.provenance_label.clone())
-        .unwrap_or_else(|| ctx.session_label.clone());
-    let origin = Origin::generated_by(label);
-    let program = parse_str(&source).map_err(|err| raise(ctx, err))?;
-    let blueprints =
-        compile_with_provenance(&program, origin.clone()).map_err(|err| raise(ctx, err))?;
-    let blueprint = blueprints
-        .into_iter()
-        .find(|b| b.graph_id == name)
-        .ok_or_else(|| {
-            invalid(
-                ctx,
-                format!("graph_define: source has no graph named `{name}`"),
-            )
-        })?;
-
-    // The draft is about to be recorded successfully; consume a slot now
-    // (re-checking the limit under the same lock to guard against a
-    // concurrent `graph_define` racing between the check above and here).
-    {
-        let mut counters = ctx.counters.lock().expect("counters poisoned");
-        if counters.graph_def >= ctx.policy.max_graph_definitions {
-            return Err(raise(
-                ctx,
-                TinyAgentsError::LimitExceeded(format!(
-                    "graph definition limit ({}) exceeded",
-                    ctx.policy.max_graph_definitions
-                )),
-            ));
-        }
-        counters.graph_def += 1;
-    }
-
-    let handle = GraphBlueprintHandle {
-        name: blueprint.graph_id.clone(),
-        source,
-        blueprint: blueprint.clone(),
-        origin,
-        compiled: false,
-        requires_review: ctx.policy.generated_graphs_require_review,
-    };
-    ctx.drafts
-        .lock()
-        .expect("drafts poisoned")
-        .insert(handle.name.clone(), handle.clone());
-    record(
-        ctx,
-        ReplCallKind::Graph,
-        "graph_define",
-        json!({ "name": handle.name }),
-        Duration::default(),
-    );
-    Ok(draft_descriptor(&handle))
-}
-
-/// Builds the script-visible descriptor map for a graph draft (carrying its
-/// name, node count, and compile/review status). The opaque
-/// [`GraphBlueprintHandle`] itself lives host-side in `ctx.drafts`.
-fn draft_descriptor(handle: &GraphBlueprintHandle) -> Dynamic {
-    let mut map = Map::new();
-    map.insert("name".into(), Dynamic::from(handle.name.clone()));
-    map.insert(
-        "nodes".into(),
-        Dynamic::from(handle.blueprint.nodes.len() as i64),
-    );
-    map.insert("compiled".into(), Dynamic::from(handle.compiled));
-    map.insert(
-        "requires_review".into(),
-        Dynamic::from(handle.requires_review),
-    );
-    Dynamic::from_map(map)
-}
-
-/// Looks up a graph draft by the `name` field of a descriptor map.
-fn lookup_draft<State: Send + Sync>(
-    ctx: &HostContext<State>,
-    descriptor: &Map,
-    func: &str,
-) -> Result<GraphBlueprintHandle, Box<EvalAltResult>> {
-    let name = map_str(descriptor, "name")
-        .ok_or_else(|| invalid(ctx, format!("{func}: descriptor is missing `name`")))?;
-    ctx.drafts
-        .lock()
-        .expect("drafts poisoned")
-        .get(&name)
-        .cloned()
-        .ok_or_else(|| invalid(ctx, format!("{func}: no graph draft named `{name}`")))
-}
-
-fn graph_validate_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    descriptor: &Map,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let handle = lookup_draft(ctx, descriptor, "graph_validate")?;
-    let program = parse_str(&handle.source).map_err(|err| raise(ctx, err))?;
-    let diagnostics = Resolver::from_registry(&*ctx.registry).resolve_program(&program);
-    let array: Array = diagnostics
-        .iter()
-        .map(|d| Dynamic::from(d.message.clone()))
-        .collect();
-    Ok(Dynamic::from_array(array))
-}
-
-fn graph_compile_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    descriptor: &Map,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let mut handle = lookup_draft(ctx, descriptor, "graph_compile")?;
-    // Bind the blueprint through the same resolver gate file-backed `.rag`
-    // source passes — generated topology is never trusted blindly.
-    Resolver::from_registry(&*ctx.registry)
-        .resolve_blueprint(&handle.blueprint)
-        .map_err(|err| raise(ctx, err))?;
-    handle.compiled = true;
-    handle.requires_review = ctx.policy.generated_graphs_require_review;
-    ctx.drafts
-        .lock()
-        .expect("drafts poisoned")
-        .insert(handle.name.clone(), handle.clone());
-    record(
-        ctx,
-        ReplCallKind::Graph,
-        "graph_compile",
-        json!({ "name": handle.name, "requires_review": handle.requires_review }),
-        Duration::default(),
-    );
-    Ok(draft_descriptor(&handle))
-}
-
-fn graph_diff_handles<State: Send + Sync>(
-    ctx: &HostContext<State>,
-    old: &Blueprint,
-    new: &Blueprint,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let diff = blueprint_diff(old, new);
-    let value = serde_json::to_value(&diff)
-        .map_err(|err| raise(ctx, TinyAgentsError::Validation(err.to_string())))?;
-    Ok(repl_value_to_dynamic(&json_to_repl_value(&value)))
-}
-
-fn graph_register_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    params: &Map,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let graph = params
-        .get("graph")
-        .and_then(|d| d.read_lock::<Map>().map(|m| m.clone()))
-        .ok_or_else(|| {
-            invalid(
-                ctx,
-                "graph_register: `graph` must be a compiled graph descriptor",
-            )
-        })?;
-    let handle = lookup_draft(ctx, &graph, "graph_register")?;
-    if !handle.compiled {
-        return Err(raise(
-            ctx,
-            TinyAgentsError::Validation(
-                "graph_register: graph must be compiled via graph_compile first".to_string(),
-            ),
-        ));
-    }
-    let review_id = map_str(params, "review_id").filter(|s| !s.is_empty());
-    if handle.requires_review && review_id.is_none() {
-        return Err(raise(
-            ctx,
-            TinyAgentsError::Validation(format!(
-                "graph_register: generated graph `{}` requires review (no review_id)",
-                handle.name
-            )),
-        ));
-    }
-    // Enforce the review gate and emit a registry intent. The compiled topology
-    // is handed to the host for installation through the registry resolver —
-    // the REPL never installs generated topology directly.
-    record(
-        ctx,
-        ReplCallKind::Graph,
-        "graph_register",
-        json!({ "name": handle.name, "review_id": review_id }),
-        Duration::default(),
-    );
-    Ok(Dynamic::from(handle.name))
-}
-
-// ── Batched implementations ─────────────────────────────────────────────────
-
-/// Extracts the object-map items of a batched argument array.
-fn batch_items<State: Send + Sync>(
-    ctx: &HostContext<State>,
-    items: &Array,
-    func: &str,
-) -> Result<Vec<Map>, Box<EvalAltResult>> {
-    items
-        .iter()
-        .map(|item| {
-            item.read_lock::<Map>()
-                .map(|m| m.clone())
-                .ok_or_else(|| invalid(ctx, format!("{func}: each item must be an object map")))
-        })
-        .collect()
-}
-
-fn model_query_batched_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    items: &Array,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    use futures::stream::{self, StreamExt};
-
-    let items = batch_items(ctx, items, "model_query_batched")?;
-    let mut prepared = Vec::with_capacity(items.len());
-    for params in &items {
-        let model_name = map_str(params, "model")
-            .ok_or_else(|| invalid(ctx, "model_query_batched: missing `model`"))?;
-        bump_model(ctx)?;
-        let model = ctx
-            .registry
-            .model(&model_name)
-            .ok_or_else(|| raise(ctx, TinyAgentsError::ModelNotFound(model_name.clone())))?;
-        let request = build_model_request(&model_name, params);
-        let structured = map_bool(params, "structured").unwrap_or(false);
-        prepared.push((model_name, model, request, structured));
-    }
-
-    let concurrency = ctx.policy.max_concurrency.max(1);
-    let results: Vec<Result<ModelBatchItem, TinyAgentsError>> =
-        bridge_block_on_raw(ctx.buffers.deadline(), async {
-            stream::iter(prepared.iter().map(|(name, model, request, structured)| {
-                let name = name.clone();
-                let structured = *structured;
-                async move {
-                    let start = Instant::now();
-                    let response = model.invoke(&ctx.state, request.clone()).await?;
-                    let finish_reason = response.finish_reason.clone();
-                    let text = Message::Assistant(response.message).text();
-                    Ok((name, text, finish_reason, structured, start.elapsed()))
-                }
-            }))
-            .buffered(concurrency)
-            .collect()
-            .await
-        })
-        .map_err(|err| raise(ctx, err))?;
-
-    let mut out = Array::with_capacity(results.len());
-    for result in results {
-        let (name, text, finish_reason, structured, elapsed) =
-            result.map_err(|err| raise(ctx, err))?;
-        record(
-            ctx,
-            ReplCallKind::Model,
-            &name,
-            json!({ "chars": text.len() }),
-            elapsed,
-        );
-        out.push(model_value(text, finish_reason, structured));
-    }
-    Ok(Dynamic::from_array(out))
-}
-
-fn tool_call_batched_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    items: &Array,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    use futures::stream::{self, StreamExt};
-
-    let items = batch_items(ctx, items, "tool_call_batched")?;
-    let mut prepared = Vec::with_capacity(items.len());
-    for params in &items {
-        let tool_name = map_str(params, "tool")
-            .ok_or_else(|| invalid(ctx, "tool_call_batched: missing `tool`"))?;
-        bump_tool(ctx)?;
-        let tool = ctx
-            .registry
-            .tool(&tool_name)
-            .ok_or_else(|| raise(ctx, TinyAgentsError::ToolNotFound(tool_name.clone())))?;
-        let arguments = map_json(params, "arguments").unwrap_or(Value::Null);
-        prepared.push((tool_name, tool, arguments));
-    }
-
-    let concurrency = ctx.policy.max_concurrency.max(1);
-    let results: Vec<
-        Result<(String, crate::harness::tool::ToolResult, Duration), TinyAgentsError>,
-    > = bridge_block_on_raw(ctx.buffers.deadline(), async {
-        stream::iter(prepared.iter().map(|(name, tool, arguments)| {
-            let name = name.clone();
-            let call = ToolCall {
-                id: new_call_id().as_str().to_string(),
-                name: name.clone(),
-                arguments: arguments.clone(),
-            };
-            async move {
-                let start = Instant::now();
-                let result = tool.call(&ctx.state, call).await?;
-                Ok((name, result, start.elapsed()))
-            }
-        }))
-        .buffered(concurrency)
-        .collect()
-        .await
-    })
-    .map_err(|err| raise(ctx, err))?;
-
-    // Each item's own tool-reported error is surfaced per item, matching the
-    // single-call path's behavior for that one call, rather than aborting the
-    // whole batch and discarding every other item's already-computed result —
-    // a batch of N independent tool calls should not lose N-1 successes
-    // because item N/2 failed. A `bridge_block_on_raw`/transport failure
-    // above (a harness-level failure, not a tool-reported one) still aborts
-    // the whole batch, since no results exist to preserve in that case.
-    let mut out = Array::with_capacity(results.len());
-    for result in results {
-        let (name, tool_result, elapsed) = result.map_err(|err| raise(ctx, err))?;
-        record(
-            ctx,
-            ReplCallKind::Tool,
-            &name,
-            json!({ "chars": tool_result.content.len() }),
-            elapsed,
-        );
-        match tool_result.error {
-            Some(error) => {
-                let mut map = Map::new();
-                map.insert("ok".into(), Dynamic::from(false));
-                map.insert("error".into(), Dynamic::from(error));
-                out.push(Dynamic::from_map(map));
-            }
-            None => {
-                let mut map = Map::new();
-                map.insert("ok".into(), Dynamic::from(true));
-                map.insert("content".into(), Dynamic::from(tool_result.content));
-                out.push(Dynamic::from_map(map));
-            }
-        }
-    }
-    Ok(Dynamic::from_array(out))
-}
-
-fn agent_query_batched_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    items: &Array,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    use crate::graph::subagent_node::SubAgentInput;
-    use futures::stream::{self, StreamExt};
-
-    let items = batch_items(ctx, items, "agent_query_batched")?;
-    let mut prepared = Vec::with_capacity(items.len());
-    for params in &items {
-        let agent_name = map_str(params, "agent")
-            .ok_or_else(|| invalid(ctx, "agent_query_batched: missing `agent`"))?;
-        bump_agent(ctx)?;
-        check_depth(ctx)?;
-        let agent = ctx.registry.agent(&agent_name).ok_or_else(|| {
-            raise(
-                ctx,
-                TinyAgentsError::Capability(format!("agent `{agent_name}` is not registered")),
-            )
-        })?;
-        let prompt = map_str(params, "prompt")
-            .or_else(|| map_str(params, "input"))
-            .unwrap_or_default();
-        let mut input = SubAgentInput::prompt(prompt);
-        if let Some(data) = map_json(params, "input") {
-            input = input.with_data(data);
-        }
-        prepared.push((agent_name, agent, input));
-    }
-
-    let concurrency = ctx.policy.max_concurrency.max(1);
-    let results: Vec<Result<AgentBatchItem, TinyAgentsError>> =
-        bridge_block_on_raw(ctx.buffers.deadline(), async {
-            stream::iter(prepared.iter().map(|(name, agent, input)| {
-                let name = name.clone();
-                async move {
-                    let start = Instant::now();
-                    let output = agent.run(input.clone(), ctx.events.clone()).await?;
-                    Ok((name, output.text, start.elapsed()))
-                }
-            }))
-            .buffered(concurrency)
-            .collect()
-            .await
-        })
-        .map_err(|err| raise(ctx, err))?;
-
-    let mut out = Array::with_capacity(results.len());
-    for result in results {
-        let (name, text, elapsed) = result.map_err(|err| raise(ctx, err))?;
-        record(ctx, ReplCallKind::Agent, &name, json!({}), elapsed);
-        out.push(Dynamic::from(text));
-    }
-    Ok(Dynamic::from_array(out))
-}
-
-fn graph_run_batched_impl<State: Send + Sync + 'static>(
-    ctx: &HostContext<State>,
-    items: &Array,
-) -> Result<Dynamic, Box<EvalAltResult>> {
-    let items = batch_items(ctx, items, "graph_run_batched")?;
-    let mut out = Array::with_capacity(items.len());
-    for params in &items {
-        out.push(graph_run_impl(ctx, params)?);
-    }
-    Ok(Dynamic::from_array(out))
-}
+use authoring::*;
+use batched::*;
+use capabilities::*;
 
 // ── Engine construction ─────────────────────────────────────────────────────
